+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+
 ### Fixed
 - fix #2: config.json was not read correct
+- fix #3: Install pyfemm on windows machines in case of not installed pyfemm
 
 ## [0.1.0] - 2021-07-28
 ### Added